Projects/App_010_001/requirements.txt
Python==3.10.12
matplotlib==3.5.1
numpy==1.24.4
pickleshare==0.7.5
pillow==9.0.1
rdkit==2023.3.3
scikit-learn==1.0.2
streamlit==1.26.0